Fix enterprise banner duplication in sso forms
