+++ packages/frontend/editor-ui/src/features/settings/sso/views/SettingsSso.vue
@@ -1,16 +1,18 @@
 <script lang="ts" setup>
 import { useDocumentTitle } from '@/app/composables/useDocumentTitle';
+import { usePageRedirectionHelper } from '@/app/composables/usePageRedirectionHelper';
 import { useSSOStore, SupportedProtocols, type SupportedProtocolType } from '../sso.store';
 import { useI18n } from '@n8n/i18n';
 import { computed, onMounted, ref } from 'vue';
 
-import { N8nHeading, N8nInfoTip, N8nOption, N8nSelect } from '@n8n/design-system';
+import { N8nActionBox, N8nHeading, N8nInfoTip, N8nOption, N8nSelect } from '@n8n/design-system';
 import SamlSettingsForm from '../components/SamlSettingsForm.vue';
 import OidcSettingsForm from '../components/OidcSettingsForm.vue';
 
 const i18n = useI18n();
 const ssoStore = useSSOStore();
 const documentTitle = useDocumentTitle();
+const pageRedirectionHelper = usePageRedirectionHelper();
 
 const options = computed(() => {
 	return [
@@ -31,6 +33,10 @@ const authProtocol = ref<SupportedProtocolType>(SupportedProtocols.SAML);
 function onAuthProtocolUpdated(value: SupportedProtocolType) {
 	authProtocol.value = value;
 }
+
+const goToUpgrade = () => {
+	void pageRedirectionHelper.goToUpgrade('sso', 'upgrade-sso');
+};
 
 onMounted(() => {
 	documentTitle.set(i18n.baseText('settings.sso.title'));
@@ -75,12 +81,30 @@ onMounted(() => {
 				</N8nSelect>
 			</div>
 		</div>
-		<div v-if="authProtocol === SupportedProtocols.SAML">
+		<div
+			v-if="ssoStore.isEnterpriseSamlEnabled && authProtocol === SupportedProtocols.SAML"
+			data-test-id="sso-content-licensed"
+		>
 			<SamlSettingsForm />
 		</div>
-		<div v-if="authProtocol === SupportedProtocols.OIDC">
+		<div
+			v-if="ssoStore.isEnterpriseOidcEnabled && authProtocol === SupportedProtocols.OIDC"
+			data-test-id="sso-content-licensed"
+		>
 			<OidcSettingsForm />
 		</div>
+		<N8nActionBox
+			v-if="!ssoStore.isEnterpriseSamlEnabled && !ssoStore.isEnterpriseOidcEnabled"
+			data-test-id="sso-content-unlicensed"
+			:class="$style.actionBox"
+			:description="i18n.baseText('settings.sso.actionBox.description')"
+			:button-text="i18n.baseText('settings.sso.actionBox.buttonText')"
+			@click:button="goToUpgrade"
+		>
+			<template #heading>
+				<span>{{ i18n.baseText('settings.sso.actionBox.title') }}</span>
+			</template>
+		</N8nActionBox>
 	</div>
 </template>
 
@@ -89,5 +113,9 @@ onMounted(() => {
 <style lang="scss" module>
 .heading {
 	margin-bottom: var(--spacing--sm);
+}
+
+.actionBox {
+	margin-top: var(--spacing--lg);
 }
 </style>
